openMP and pthread source codes debug, openMP convergence

+++ lab1_omp.h
@@ -39,7 +39,7 @@ void kmeans_omp(int num_threads,
 				int K,
 				int* data_points,
 				int** data_point_cluster,
-				int** centroids,
+				float** centroids,
 				int* num_iterations
 				);
